Date: Use more fine-grained types.

+++ src/Date.h
@@ -8,12 +8,12 @@
 #include "Exception.h"
 
 #define Date_Define() \
-	short year;       \
-	short month;      \
-	short day;        \
-	short hour;       \
-	short minute;     \
-	short second
+	uint16_t year;    \
+	uint8_t  month;   \
+	uint8_t  day;     \
+	uint8_t  hour;    \
+	uint8_t  minute;  \
+	uint8_t  second
 
 typedef struct {
 	Date_Define();
